feat: rename of packages
